Changed to newer AWS Lib

+++ Amazon.S3/Model/S3ListMultipartUploadsRequest.m
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2012 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2013 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -62,13 +62,13 @@
     return urlRequest;
 }
 
--(void)setMaxUploads:(int)max
+-(void)setMaxUploads:(NSInteger)max
 {
     maxUploads      = max;
     maxUploadsIsSet = YES;
 }
 
--(int)maxUploads
+-(NSInteger)maxUploads
 {
     return maxUploads;
 }
