+++ include/S3/S3ListMultipartUploadsRequest.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2012 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2013 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -19,16 +19,15 @@
 
 /** Contains the parameters used for the listMultipartUploads operation.
  *
- * \ingroup S3
  */
 @interface S3ListMultipartUploadsRequest:S3Request {
-    NSString *delimiter;
-    NSString *keyMarker;
-    NSString *prefix;
-    NSString *uploadIdMarker;
+    NSString  *delimiter;
+    NSString  *keyMarker;
+    NSString  *prefix;
+    NSString  *uploadIdMarker;
 
-    int      maxUploads;
-    bool     maxUploadsIsSet;
+    NSInteger maxUploads;
+    bool      maxUploadsIsSet;
 }
 
 
@@ -72,6 +71,6 @@
 @property (nonatomic, retain) NSString *uploadIdMarker;
 
 /** Gets and sets the maximum number of multipart uploads to return in the response. */
-@property (nonatomic) int maxUploads;
+@property (nonatomic) NSInteger maxUploads;
 
 @end
