+++ include/S3/S3ListPartsRequest.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2012 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2013 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -19,12 +19,11 @@
 
 /** Contains the parameters used for the listParts operation.
  *
- * \ingroup S3
  */
 @interface S3ListPartsRequest:S3Request {
-    NSString *uploadId;
-    int      maxParts;
-    int      partNumberMarker;
+    NSString  *uploadId;
+    NSInteger maxParts;
+    NSInteger partNumberMarker;
 }
 
 /** The uploadId of the multipart upload to list */
@@ -34,10 +33,10 @@
  *
  * The default value is 1000.
  */
-@property (nonatomic) int maxParts;
+@property (nonatomic) NSInteger maxParts;
 
 /** The part after which to start the listing. */
-@property (nonatomic) int partNumberMarker;
+@property (nonatomic) NSInteger partNumberMarker;
 
 /** Initializes the instance with the bucket, key, and uploadId from an S3MultipartUpload instance returned in an S3MultipartUploadResponse. */
 -(id)initWithMultipartUpload:(S3MultipartUpload *)multipartUpload;
