+++ include/S3/S3ListPartsResult.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2012 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2013 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -18,7 +18,6 @@
 
 /** The S3ListPartsResult contains all the information about the listParts method.
  *
- * \ingroup s3
  */
 @interface S3ListPartsResult:NSObject {
     NSString       *bucket;
@@ -29,9 +28,9 @@
     S3Owner        *owner;
     S3Owner        *initiator;
 
-    int            partNumberMarker;
-    int            nextPartNumberMarker;
-    int            maxParts;
+    NSInteger      partNumberMarker;
+    NSInteger      nextPartNumberMarker;
+    NSInteger      maxParts;
 
     bool           isTruncated;
 
@@ -57,16 +56,16 @@
 @property (nonatomic, retain) S3Owner *initiator;
 
 /** The part number after which listing begins. */
-@property (nonatomic) int partNumberMarker;
+@property (nonatomic) NSInteger partNumberMarker;
 
 /** When a list is truncated, specifies the last part that should be skipped
  * over to resume listing. Use this value for the partNumberMarker
  * request property in a subsequent request.
  */
-@property (nonatomic) int nextPartNumberMarker;
+@property (nonatomic) NSInteger nextPartNumberMarker;
 
 /** The maxParts which was supplied in the request. */
-@property (nonatomic) int maxParts;
+@property (nonatomic) NSInteger maxParts;
 
 /** Indicates whether the returned list of parts is truncated. A value true
  * indicates the list was truncated. A list may be truncated if the number of
